Restaurant Verification Update

+++ src/main/java/com/example/deliciousBee/controller/report/ReportController.java
@@ -4,15 +4,19 @@ import java.time.LocalDate;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 
 import com.example.deliciousBee.dto.report.ReportDto;
 import com.example.deliciousBee.dto.report.RestaurantVerificationDto;
+import com.example.deliciousBee.model.board.Restaurant;
+import com.example.deliciousBee.model.board.VerificationStatus;
 import com.example.deliciousBee.model.member.BeeMember;
 import com.example.deliciousBee.model.report.Report;
 import com.example.deliciousBee.repository.ReportRepository;
 import com.example.deliciousBee.service.report.ReportService;
 import com.example.deliciousBee.service.restaurant.RestaurantService;
+import com.example.deliciousBee.service.review.ReviewService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
@@ -30,8 +34,7 @@ public class ReportController {
 
 	private final RestaurantService restaurantService;
 	private final ReportService reportService;
-	private final ReportRepository reportRepository;
-
+	private final ReviewService reviewService;
 
 	@GetMapping("/admin")
 	public String adminPage() {
@@ -93,12 +96,38 @@ public class ReportController {
 	public ResponseEntity<Map<String, Object>> deleteReport(@PathVariable("reportId") Long reportId) {
 		Map<String, Object> response = new HashMap<>();
 		try {
-			reportService.deleteReport(reportId);
-			response.put("success", true);
+			Optional<Report> reportOptional = reportService.getReportById(reportId);
+			if (reportOptional.isPresent()) {
+				Report report = reportOptional.get();
+				Long reviewId = report.getReview().getId(); // Report 엔티티에서 reviewId 가져오기
+				reportService.deleteReport(reportId);
+				reviewService.deleteReview(reviewId);
+				response.put("success", true);
+			} else {
+				response.put("success", false);
+				response.put("message", "Report not found");
+			}
 		} catch (Exception e) {
 			response.put("success", false);
 			response.put("message", "서버 오류가 발생했습니다.");
 			log.error("Report deleting error", e);
+		}
+		return ResponseEntity.ok(response);
+	}
+
+
+
+	@PostMapping("/admin/restaurant/approve/{restaurantId}")
+	@ResponseBody
+	public ResponseEntity<Map<String,Object>> approveRestaurant(@PathVariable("restaurantId") Long restaurantId) {
+		Map<String, Object> response = new HashMap<>();
+		try {
+			Restaurant restaurant = restaurantService.findRestaurant(restaurantId);
+			restaurantService.updateApprove(restaurant);
+		} catch (Exception e) {
+			response.put("success", false);
+			response.put("message", "서버 오류가 발생했습니다.");
+			log.error("Report getting error", e);
 		}
 		return ResponseEntity.ok(response);
 	}
@@ -120,7 +149,6 @@ public class ReportController {
 		}
 		return ResponseEntity.ok(response);
 	}
-
